refactor: 페에지 폴더 구조를 변경(#9)
