+++ src/pages/todo/TodoPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Grid } from '@mui/material';
 import { Outlet } from 'react-router-dom';
-import TodoListContainer from '../components/todo/container/TodolistContainer';
+import TodoListContainer from '../../components/todo/container/TodolistContainer';
 
 type Props = {};
